+++ CHANGELOG.md
@@ -35,6 +35,8 @@
 - Each Neo4j query gets logged when the log level is `info`. It will help in debugging slow queries.
 - Duration of keeping expired sessions in the DB can be configured via `sirko.engine.stale_session_in` config setting.
     To get more details about this setting check the config/sirko.conf file.
+- A new `sirko.engine.confidence_threshold` setting is introduced to define a confidence threshold which must be met
+    to prerender a predicted page. Use this setting to reduce load on your backend if you get some because of the prerendering.
 
 ## v0.0.2 (27 March 2017)